Corregge un mismatch di tipo in SoundEngine (Long vs Int) che rompeva la build
inviluppoNota riceveva durataNota (Long, i campioni per nota) ma la firma si
aspettava un Int: la CI del redesign falliva in compilazione Kotlin
(compileReleaseKotlin). Cambio il parametro a Long, coerente col resto
dell'aritmetica sui campioni per nota.

+++ game/src/main/java/it/example/frattalogic/audio/SoundEngine.kt
@@ -198,7 +198,7 @@ class SoundEngine {
         return if (t < attacco) t / attacco else exp(-3.0 * (t - attacco) / durata)
     }
 
-    private fun inviluppoNota(posizione: Int, durata: Int): Double {
+    private fun inviluppoNota(posizione: Int, durata: Long): Double {
         val attacco = (durata * 0.06).coerceAtLeast(1.0)
         return when {
             posizione < attacco -> posizione / attacco
